fix(shadows): add castsShadows

+++ game/src/layer/mazelayer.cpp
@@ -4,10 +4,9 @@
 #include "scene/pointlight.hpp"
 #include "sponge.hpp"
 #include <glm/ext/matrix_transform.hpp>
-#include <glm/gtc/matrix_transform.hpp> // For ortho, perspective, lookAt
-#include <memory> // For std::make_unique
+#include <glm/gtc/matrix_transform.hpp>  // For ortho, perspective, lookAt
 #include <array>
-#include "sponge/platform/opengl/scene/shadowmap.hpp" // For ShadowMap class
+#include <memory>  // For std::make_unique
 
 namespace {
 constexpr auto keyboardSpeed = .075F;
@@ -85,7 +84,9 @@ void MazeLayer::onAttach() {
     shader->setFloat("ambientStrength", ambientStrength);
     shader->unbind();
 
-    // shadowMap = std::make_unique<sponge::platform::opengl::scene::ShadowMap>(); // Removed
+    // shadowMap =
+    // std::make_unique<sponge::platform::opengl::scene::ShadowMap>(); //
+    // Removed
 
     cube = std::make_unique<sponge::platform::opengl::scene::Cube>();
 
@@ -151,7 +152,8 @@ void MazeLayer::setNumLights(const int32_t val) {
 
     for (int32_t i = 0; i < numLights; ++i) {
         if (!pointLights[i].shadowMap) {
-            pointLights[i].shadowMap = std::make_unique<sponge::platform::opengl::scene::ShadowMap>();
+            pointLights[i].shadowMap =
+                std::make_unique<sponge::platform::opengl::scene::ShadowMap>();
         }
 
         pointLights[i].color = glm::vec3(1.F);
@@ -160,29 +162,37 @@ void MazeLayer::setNumLights(const int32_t val) {
                    glm::vec3(0.F, 1.F, 0.F)) *
             glm::vec4(-1.F, 2.75F, -1.F, 1.F));
         // Update position based on translation for consistency
-        // pointLights[i].position = pointLights[i].translation; // This will be set by specific light configs
+        // pointLights[i].position = pointLights[i].translation; // This will be
+        // set by specific light configs
 
-        if (i == 0) { // Directional Light
+        if (i == 0) {  // Directional Light
             pointLights[i].type = LightType::DIRECTIONAL;
-            pointLights[i].position = glm::vec3(0.f, 7.f, 7.f); // Position of the light volume/camera for ortho proj
-            pointLights[i].direction = glm::normalize(glm::vec3(0.0f, -0.5f, -0.5f));
+            pointLights[i].position = glm::vec3(
+                0.f, 7.f,
+                7.f);  // Position of the light volume/camera for ortho proj
+            pointLights[i].direction =
+                glm::normalize(glm::vec3(0.0f, -0.5f, -0.5f));
             pointLights[i].castsShadows = 1;
             pointLights[i].color = glm::vec3(0.8f);
-            // fov and aspectRatio are not typically used for directional lights in the same way,
-            // but can be set to default values if PointLight struct requires them.
-            // The ortho projection matrix will define its volume.
-            pointLights[i].fov = glm::radians(45.0f); // Default
-            pointLights[i].aspectRatio = 1.0f;      // Default
-        } else if (i == 1) { // Spot Light
+            // fov and aspectRatio are not typically used for directional lights
+            // in the same way, but can be set to default values if PointLight
+            // struct requires them. The ortho projection matrix will define its
+            // volume.
+            pointLights[i].fov = glm::radians(45.0f);  // Default
+            pointLights[i].aspectRatio = 1.0f;         // Default
+        } else if (i == 1) {                           // Spot Light
             pointLights[i].type = LightType::SPOT;
             pointLights[i].position = glm::vec3(-3.f, 4.f, -3.f);
-            pointLights[i].translation = pointLights[i].position; // Keep translation consistent
-            pointLights[i].direction = glm::normalize(glm::vec3(1.0f, -1.0f, 1.0f));
+            pointLights[i].translation =
+                pointLights[i].position;  // Keep translation consistent
+            pointLights[i].direction =
+                glm::normalize(glm::vec3(1.0f, -1.0f, 1.0f));
             pointLights[i].fov = glm::radians(30.0f);
             pointLights[i].aspectRatio = 1.0f;
             pointLights[i].castsShadows = 1;
             pointLights[i].color = glm::vec3(1.0f, 1.0f, 0.8f);
-        } else { // Other lights (e.g., Point, not casting shadows for this test)
+        } else {  // Other lights (e.g., Point, not casting shadows for this
+                  // test)
             pointLights[i].type = LightType::POINT;
             // Default position based on loop, or set explicitly if needed
             pointLights[i].position = glm::vec3(
@@ -190,14 +200,17 @@ void MazeLayer::setNumLights(const int32_t val) {
                        glm::vec3(0.F, 1.F, 0.F)) *
                 glm::vec4(-1.F, 2.75F, -1.F, 1.F));
             pointLights[i].translation = pointLights[i].position;
-            pointLights[i].direction = glm::normalize(glm::vec3(0.0f, -1.0f, 0.0f)); // Default
+            pointLights[i].direction =
+                glm::normalize(glm::vec3(0.0f, -1.0f, 0.0f));  // Default
             pointLights[i].fov = glm::radians(90.0f);
             pointLights[i].aspectRatio = 1.0f;
-            pointLights[i].castsShadows = 0; // Not casting shadows for this test
-            pointLights[i].color = glm::vec3(0.5f); // Dimmer color
+            pointLights[i].castsShadows =
+                0;  // Not casting shadows for this test
+            pointLights[i].color = glm::vec3(0.5f);  // Dimmer color
         }
 
-        // Common properties like attenuation can be set outside or after specific configs
+        // Common properties like attenuation can be set outside or after
+        // specific configs
         pointLights[i].setAttenuationFromIndex(attenuationIndex);
     }
 
@@ -263,47 +276,66 @@ bool MazeLayer::onWindowResize(
 
 void MazeLayer::renderGameObjects() const {
     constexpr float nearPlane = 1.0f;
-    constexpr float farPlane = 25.0f; // Adjusted far plane for potentially larger scenes or ortho views
+    constexpr float farPlane = 25.0f;  // Adjusted far plane for potentially
+                                       // larger scenes or ortho views
 
-    auto shadowShader = ResourceManager::getShader(sponge::platform::opengl::scene::ShadowMap::getShaderName());
+    auto shadowShader = ResourceManager::getShader(
+        sponge::platform::opengl::scene::ShadowMap::getShaderName());
 
     // Shadow Pass: Render scene from each light's perspective
-    glViewport(0, 0, sponge::platform::opengl::scene::ShadowMap::SHADOW_WIDTH, sponge::platform::opengl::scene::ShadowMap::SHADOW_HEIGHT); // Set viewport to shadow map size
+    glViewport(0, 0, sponge::platform::opengl::scene::ShadowMap::SHADOW_WIDTH,
+               sponge::platform::opengl::scene::ShadowMap::
+                   SHADOW_HEIGHT);  // Set viewport to shadow map size
 
     for (int i = 0; i < numLights; ++i) {
         PointLight& light = pointLights[i];
-        if (!light.shadowMap) continue;
+        if (!light.shadowMap)
+            continue;
 
         light.shadowMap->bind();
-        glClear(GL_DEPTH_BUFFER_BIT); // Clear depth buffer for this light's shadow map
+        glClear(GL_DEPTH_BUFFER_BIT);  // Clear depth buffer for this light's
+                                       // shadow map
 
         // Calculate View Matrix
         if (light.type == LightType::DIRECTIONAL) {
-            // Directional light looks from its position towards position + direction
-            light.lightViewMatrix = glm::lookAt(light.position, light.position + light.direction, glm::vec3(0.0f, 1.0f, 0.0f));
-        } else { // SPOT and POINT (simplified for now, point lights would ideally use cubemaps)
-            light.lightViewMatrix = glm::lookAt(light.position, light.position + light.direction, glm::vec3(0.0f, 1.0f, 0.0f));
+            // Directional light looks from its position towards position +
+            // direction
+            light.lightViewMatrix =
+                glm::lookAt(light.position, light.position + light.direction,
+                            glm::vec3(0.0f, 1.0f, 0.0f));
+        } else {  // SPOT and POINT (simplified for now, point lights would
+                  // ideally use cubemaps)
+            light.lightViewMatrix =
+                glm::lookAt(light.position, light.position + light.direction,
+                            glm::vec3(0.0f, 1.0f, 0.0f));
         }
 
         // Calculate Projection Matrix
         if (light.type == LightType::DIRECTIONAL) {
             // Orthographic projection for directional lights
-            light.lightProjectionMatrix = glm::ortho(-15.0f, 15.0f, -15.0f, 15.0f, nearPlane, farPlane);
-        } else { // SPOT and POINT (perspective projection)
-            // Ensure FOV is reasonable, e.g. not zero, for perspective projection
-            float currentFov = (light.fov > glm::radians(0.01f)) ? light.fov : glm::radians(45.0f);
-            light.lightProjectionMatrix = glm::perspective(currentFov, light.aspectRatio, nearPlane, farPlane);
+            light.lightProjectionMatrix =
+                glm::ortho(-15.0f, 15.0f, -15.0f, 15.0f, nearPlane, farPlane);
+        } else {  // SPOT and POINT (perspective projection)
+            // Ensure FOV is reasonable, e.g. not zero, for perspective
+            // projection
+            float currentFov = (light.fov > glm::radians(0.01f))
+                                   ? light.fov
+                                   : glm::radians(45.0f);
+            light.lightProjectionMatrix = glm::perspective(
+                currentFov, light.aspectRatio, nearPlane, farPlane);
         }
 
-        glm::mat4 currentLightSpaceMatrix = light.lightProjectionMatrix * light.lightViewMatrix;
+        glm::mat4 currentLightSpaceMatrix =
+            light.lightProjectionMatrix * light.lightViewMatrix;
 
         shadowShader->bind();
         shadowShader->setMat4("lightSpaceMatrix", currentLightSpaceMatrix);
 
         for (const auto& gameObject : gameObjects) {
             const auto model = glm::scale(
-                glm::rotate(glm::translate(glm::mat4(1.0f), gameObject.translation),
-                            gameObject.rotation.angle, gameObject.rotation.axis),
+                glm::rotate(
+                    glm::translate(glm::mat4(1.0f), gameObject.translation),
+                    gameObject.rotation.angle, gameObject.rotation.axis),
                 gameObject.scale);
             shadowShader->setMat4("model", model);
             ResourceManager::getModel(gameObject.name)->render(shadowShader);
@@ -312,28 +344,33 @@ void MazeLayer::renderGameObjects() const {
     }
     shadowShader->unbind();
 
-    // Update shader light uniforms AFTER shadow maps have been generated and matrices calculated
+    // Update shader light uniforms AFTER shadow maps have been generated and
+    // matrices calculated
     updateShaderLights();
 
     // Main Scene Rendering Pass
     glViewport(0, 0, Maze::get().getWidth(), Maze::get().getHeight());
-    glClear(GL_COLOR_BUFFER_BIT | GL_DEPTH_BUFFER_BIT); // Clear default framebuffer
+    glClear(GL_COLOR_BUFFER_BIT |
+            GL_DEPTH_BUFFER_BIT);  // Clear default framebuffer
 
-    auto pbrShader = ResourceManager::getShader(sponge::platform::opengl::scene::Mesh::getShaderName());
+    auto pbrShader = ResourceManager::getShader(
+        sponge::platform::opengl::scene::Mesh::getShaderName());
     pbrShader->bind();
 
     // Activate and Bind all necessary shadow map textures
-    for (int i = 0; i < numLights; ++i) {
+    for (uint8_t i = 0; i < numLights; ++i) {
         if (pointLights[i].castsShadows == 1 && pointLights[i].shadowMap) {
             // Texture unit 0 is usually for diffuse. Shadow maps start from 1.
-            // This texture unit (i + 1) must match what's set in updateShaderLights for pointLights[i].shadowMapSampler
+            // This texture unit (i + 1) must match what's set in
+            // updateShaderLights for pointLights[i].shadowMapSampler
             pointLights[i].shadowMap->activateAndBindDepthMap(i + 1);
         }
     }
 
     pbrShader->setFloat3("viewPos", camera->getPosition());
-    // Other PBR uniforms (metallic, roughness, ao, ambientStrength) are set in onAttach or setters
-    // Light uniforms (including lightSpaceMatrix and shadowMapSampler) are set by updateShaderLights()
+    // Other PBR uniforms (metallic, roughness, ao, ambientStrength) are set in
+    // onAttach or setters Light uniforms (including lightSpaceMatrix and
+    // shadowMapSampler) are set by updateShaderLights()
 
     for (const auto& gameObject : gameObjects) {
         const auto model = glm::scale(
@@ -403,25 +440,36 @@ void MazeLayer::updateShaderLights() const {
 
     for (int32_t i = 0; i < numLights; ++i) {
         // Ensure position is updated if translation changes.
-        // This was done in setNumLights, if lights are dynamic, this might need to be updated each frame.
-        // pointLights[i].position = glm::vec4(pointLights[i].translation, 1.F); // Already vec3, ensure consistency.
-                                                                               // PointLight::position is already glm::vec3.
+        // This was done in setNumLights, if lights are dynamic, this might need
+        // to be updated each frame. pointLights[i].position =
+        // glm::vec4(pointLights[i].translation, 1.F); // Already vec3, ensure
+        // consistency. PointLight::position is already glm::vec3.
 
         pbrShader->setFloat3("pointLights[" + std::to_string(i) + "].position",
-                          pointLights[i].position);
-        pbrShader->setFloat3("pointLights[" + std::to_string(i) + "].attenuation",
-                          pointLights[i].getAttenuation());
+                             pointLights[i].position);
+        pbrShader->setFloat3(
+            "pointLights[" + std::to_string(i) + "].attenuation",
+            pointLights[i].getAttenuation());
         pbrShader->setFloat3("pointLights[" + std::to_string(i) + "].color",
-                          pointLights[i].color);
+                             pointLights[i].color);
 
-        // Calculate LightSpaceMatrix for the shader (using matrices computed in shadow pass)
-        glm::mat4 lightSpaceMatrixForShader = pointLights[i].lightProjectionMatrix * pointLights[i].lightViewMatrix;
-        pbrShader->setMat4("pointLights[" + std::to_string(i) + "].lightSpaceMatrix", lightSpaceMatrixForShader);
+        // Calculate LightSpaceMatrix for the shader (using matrices computed in
+        // shadow pass)
+        glm::mat4 lightSpaceMatrixForShader =
+            pointLights[i].lightProjectionMatrix *
+            pointLights[i].lightViewMatrix;
+        pbrShader->setMat4(
+            "pointLights[" + std::to_string(i) + "].lightSpaceMatrix",
+            lightSpaceMatrixForShader);
 
-        // Texture units for shadow maps typically start from 1 (0 is often diffuse)
-        // This needs to match the activateAndBindDepthMap call in renderGameObjects
-        pbrShader->setInteger("pointLights[" + std::to_string(i) + "].shadowMapSampler", i + 1);
-        pbrShader->setInteger("pointLights[" + std::to_string(i) + "].castsShadows", pointLights[i].castsShadows);
+        // Texture units for shadow maps typically start from 1 (0 is often
+        // diffuse) This needs to match the activateAndBindDepthMap call in
+        // renderGameObjects
+        pbrShader->setInteger(
+            "pointLights[" + std::to_string(i) + "].shadowMapSampler", i + 1);
+        pbrShader->setInteger(
+            "pointLights[" + std::to_string(i) + "].castsShadows",
+            pointLights[i].castsShadows);
     }
 
     pbrShader->unbind();
